misc: rename query file
